pyproject.toml
[project]
name = "chatbot-fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "langgraph>=0.3.18",
    "langgraph-cli[inmem]>=0.1.79",
    "langmem>=0.0.16",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
